Passport authentication added to all sensitive routes in app
